+++ .cargo/config.toml
@@ -17,7 +17,7 @@ rustflags=["-C", "link-arg=-fuse-ld=lld"]
 # - Ubuntu, `sudo apt install lld clang`
 # - Arch, `sudo pacman -S lld clang`
 [target.x86_64-unknown-linux-gnu]
-rustflags=["-C", "linker=clang", "-C", "link-arg=-fuse-ld=lld"]
+rustflags=["-C", "linker=clang", "-C", "link-arg=-fuse-ld=lld-15"]
 
 # On MacOS, `brew install michaeleisel/zld/zld
 [target.x86_64-apple-darwin]
