+++ tests/api/main.rs
@@ -1,0 +1,3 @@
+mod health_check;
+mod helpers;
+mod subscriptions;
